code_base/crm_api/requirements.txt
psycopg2-binary
python-dotenv
sqlalchemy
pandas 
fastapi
uvicorn
pydantic
python-jose[cryptography]
passlib[bcrypt]
python-multipart
httpx 